requirements.txt

openai>=1.43.0
langchain>=0.2.12
langchain-openai>=0.1.22
python-dotenv>=1.0.1
pytest>=8.2.0
